requirements.txt
streamlit
scrapy
faiss-cpu
sentence-transformers
google-generativeai
pandas
numpy
scikit-learn
python-dotenv
transformers
tqdm
pathlib
fastapi
uvicorn